+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
-**./app/test-common.js**
-* Wrote new function 'checkNumber'
+**./test-parts/01-common_data/items/itm-common_objects.js**
+* Renamed global variables:
+	* 'testCommon' to 'commonFunctions'
+	* 'testArray' to 'arrayFunctions'
+* Removed 'commonPaths.testObject' requirement.